feat: metadata_cleaner() cleaned all file metadata
"metadata_cleaner.py" written to contain a series of useful functions.
These functions allow me to use the command line to
- read in original file from "Original Data\"
- create modified file in "Modified Data\"
By passing a folder to "folder_clean()", all of the "#_delay.txt" files
are read in, and rewritten with clean metadata.
- simple_path() builds a full filepath from the folder
- dlist_gen() produces all the relevant files from the folder
- passes file by file to metadata_cleaner() to produce new files.

This process was done on all the relevant data folders:
"2016-09-22", "2016-09-23", "2016-09-24", "2016-09-26", "2016-09-27"
"2016-09-28", "2016-10-01", "2016-10-04"
I gave file a quick visual check, and it looks okay.