Typos in data files corrected with "typo_fix()"
"typo_fix()" in "metadata_cleaner.py" fixes specific errors in the final
metadata for a file before it's reprinted into "Modified Data\".
"typo_fix(metadata)" checks for a file with a known typo, and then
replaces the bad metadata with the correct metadata.

Please enter the commit message for your changes. Lines starting